device_handle: Enable automatic kernel driver detachment
When this is enabled libusb will automatically detach the kernel
driver on an interface when claiming the interface, and attach it when
releasing the interface.

Automatic kernel driver detachment is disabled on newly opened device
handles by default.

On platforms which do not have support, this function will return
`Error::NotSupported`, and rusb will continue as if this function was
never called.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>