Use "system" calling convention for Win32 compatibility
The "system" calling convention is, usually same as extern "C". The
exception is on Win32, in which case it's "stdcall", or what you
should use to link to the Windows API itself.

One exception to this rule was "libusb_set_option" as it is a variatic
function and this must use "C" or "cdecl" calling conventions as it is
almost impossible to get it right with "stdcall", because only the
caller really knows how many arguments were passed in order to clean
them up.

This fixes linking on Win32 platforms.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>