error: Refactor code to drop deprecated call and reduce code
The `Error::description` method has been deprecated and people should
use `Display` or `to_string()` to gather the same kind of output. Doing
so, allowed the removal of `strerror` auxiliary method.

We also removed some import renames, so it has less indirection for the
code and make it easier to read.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>