Increase endpoint descriptor's lifetime
Before this commit, the lifetime of the endpoint descriptor was shorter
than the lifetime of the config descriptor, since it was tied to the
InterfaceDescriptor object. This commit amends the endpoint_descriptors
signature so that an EndpointDescriptor has the same lifetime as the
overarching ConfigDescriptor.

From my understanding, all state derived from a ConfigDescriptor has a
lifetime associated to that ConfigDescriptor. So, it's safe for the
EndpointDescriptor to adopt the ConfigDescriptor's lifetime through the
InterfaceDescriptor.

The new test shows us that the lifetimes are now all equal. This test
wouldn't compile without the signature change.