fix: endpoint number parsing error
Fixed a bug in parsing the ndpoint number from the address field in endpoint descriptor. According to the USB device specification, bits 0-3 of the address field in the endpoint descriptor represent the endpoint number. Therefore, when parsing the number from the address, it should be masked with 0xf instead of 0x7. Incorrect parsing will lead to corrupted device endpoint information.